pca-tsne-umap-plotter/requirements.txt
numpy>=1.20
matplotlib>=3.0
scikit-learn>=0.24
umap-learn>=0.5